refactor: restructure the whole project
