+++ src/index.css
@@ -1,7 +1,6 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
 @layer base {
     :root {
         --background: 0 0% 100%;
@@ -10,7 +9,7 @@
         --card-foreground: 222.2 84% 4.9%;
         --popover: 0 0% 100%;
         --popover-foreground: 222.2 84% 4.9%;
-        --primary: 222.2 47.4% 11.2%;
+        --primary: 238.7 83.5% 66.7%;
         --primary-foreground: 210 40% 98%;
         --secondary: 210 40% 96.1%;
         --secondary-foreground: 222.2 47.4% 11.2%;
@@ -58,18 +57,6 @@
         --chart-5: 340 75% 55%;
     }
 }
-
-@layer base {
-    * {
-        @apply border-border;
-    }
-
-    body {
-        @apply text-foreground;
-        background: #fbfbfb;
-    }
-}
-
 
 /*calendar style*/
